[#5]fix: 예외 구조 변경
