+++ app/web/exception_handlers.py
@@ -1,6 +1,6 @@
-from app.exception.base_custom_exception import BaseCustomException
-from app.exception.invalid_exception import InvalidException
-from app.models.error_response import ErrorResponse
+from app.route.exception.base_custom_exception import BaseCustomException
+from app.route.exception.invalid_exception import InvalidException
+from app.route.models.error_response import ErrorResponse
 
 from fastapi import FastAPI, Request
 from fastapi.responses import JSONResponse
@@ -10,23 +10,23 @@ def setup_exception_handlers(app: FastAPI):
     @app.exception_handler(InvalidException)
     async def invalid_exception_handler(request: Request, exc: InvalidException):
         response = ErrorResponse(
-            code=exc.custom_error.code,
-            detail=exc.custom_error.detail,
+            code=exc.error_enum.code,
+            detail=exc.error_enum.detail,
             result=exc.result
         )
         return JSONResponse(
-            status_code=exc.status_code,
+            status_code=exc.error_enum.status,
             content=response.to_dict()
         )
 
     @app.exception_handler(BaseCustomException)
     async def base_exception_handler(request: Request, exc: BaseCustomException):
         response = ErrorResponse(
-            code=exc.custom_error.code,
-            detail=exc.custom_error.detail,
+            code=exc.error_enum.code,
+            detail=exc.error_enum.detail,
             result=exc.result
         )
         return JSONResponse(
-            status_code=exc.status_code,
+            status_code=exc.error_enum.status,
             content=response.to_dict()
         )
